backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17
python-pptx>=1.0.2
pytest>=8.0.0
httpx>=0.27.0
